engine/requirements.txt
requests==2.19.1
Pillow==7.0.0
jsonobject==0.9.9
rx==3.0.1
opencv-python==4.1.0.25
